requirements.txt
absl-py==0.10.0
apturl==0.5.2
argon2-cffi==20.1.0
asn1crypto==0.24.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
beautifulsoup4==4.6.0
bert-pytorch==0.0.1a4
bert-tensorflow==1.0.4
bleach==3.1.5
Brlapi==0.6.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.21
dataclasses==0.7
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
future==0.18.2
gast==0.3.3
gluonnlp==0.10.0
google-auth==1.21.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.32.0
h5py==2.10.0
httplib2==0.9.2
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
JPype1==1.0.2
json5==0.9.5
jsonpatch==1.26
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.8
jupyterlab-pygments==0.1.1
jupyterlab-server==1.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
kobert==0.1.1
kobert-transformers==0.4.1
konlpy==0.5.2
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
llvmlite==0.34.0
louis==3.5.0
lxml==4.5.2
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.2
MarkupSafe==1.0
matplotlib==3.3.2
mecab-python===0.996-ko-0.9.2
mistune==0.8.4
mxnet==1.7.0.post1
mxnet-cu101==1.7.0
nbclient==0.5.0
nbconvert==6.0.2
nbformat==5.0.7
nest-asyncio==1.4.0
netifaces==0.10.4
notebook==6.1.4
numba==0.51.2
numpy==1.18.5
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
onnxruntime==1.4.0
openpyxl==3.0.5
opt-einsum==3.3.0
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.2.1
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.7.0
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
pyrsistent==0.17.3
PySocks==1.7.1
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.1
python-debian==0.1.32
pytorch-transformers==1.2.0
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
pyzmq==19.0.2
reportlab==3.4.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.43
scipy==1.4.1
scour==0.36
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.13.2
six==1.15.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.1
tensorflow-hub==0.9.0
tensorflow-text==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torchfile==0.1.0
tornado==6.0.4
traitlets==4.3.3
tweepy==3.9.0
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.10
usb-creator==0.3.3
virtualenv==15.1.0
visdom==0.1.8.9
wadllib==1.3.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
xlrd==1.2.0
zipp==3.1.0
zope.interface==4.3.2
